服务外包比赛/backend/requirements.txt
# FastAPI核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
neo4j==5.16.0

# 数据处理
pandas==2.1.4
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.6

# AI模型
transformers==4.37.0
torch==2.1.2
jieba==0.42.1

# 图形处理
networkx==3.2.1
matplotlib==3.8.2
Pillow==10.2.0

# 工具库
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
